Add onPrintDialogClose callback option
This option enables print "completion" notification, when possible.
Note that these implementation is somehow weak as we don't currently
have a way to know whether the user printed the document or not, so
the use of "print completion notification" is misleading.

Another issue to highlight is that this implementation depends upon
window's mouseover event, this means that, while the print dialog
is open, the window does not have the focus and thus, the mouseover
event will not be fired. However, this is not the case for Firefox,
because the print dialog does not removes the focus from window.

Last, the implementation is inspired in StackOverflow's thread [how
to detech window.print() finish](https://stackoverflow.com/q/18325025)
and the two responses https://stackoverflow.com/a/35430328 and
https://stackoverflow.com/a/45625229