Fix bug that stopped images from printing.
This bug caused the image to not print unless it was already loaded when `loadIframeImage()` was first called.  Otherwise any subsequent calls to `loadIframeImage()` would create a new promise and then resolve that new promise -- instead of resolving the promise that was being waited on by the `Promise.all()` returned by `loadIframeImages()`.