feat: send XMLHttpRequest as second param for onError
Currently the `onError` callback only has the `req.statusText` field,
as param.

Would be useful to have the XMLHttpRequest object instance along
with the req.statusText, so we are able to consume all the details
from the request, the response (ArrayBuffer) for instance.

When available the XMLHttpRequest instance used in the request
is sent as second param for the `onError` callback.